Переводит компоненты с данными в классы
